put remember the milk model in correctly named package
