+++ extensions/data-transfer/portability-data-transfer-rememberthemilk/src/main/java/org/dataportabilityproject/transfer/rememberthemilk/model/tasks/Frob.java
@@ -13,12 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.transfer.microsoft.model.tasks;
+package org.dataportabilityproject.transfer.rememberthemilk.model.tasks;
 
 import com.google.api.client.util.Key;
 
-public class TimelineCreateResponse extends RememberTheMilkResponse {
+/**
+ * Response from rtm.auth.getFrob
+ *
+ * <p>Example: //<?xml version='1.0' encoding='UTF-8'?><rsp
+ * stat="ok"><frob>d27f3ecf5497d7fdd79aea0ba1ebe9bad375ce7b</frob></rsp>
+ */
+public class Frob extends RememberTheMilkResponse {
 
-  @Key("timeline")
-  public String timeline;
+  @Key("frob")
+  public String frob;
+
+  @Override
+  public String toString() {
+    return "frob(stat=" + stat + ", frob=" + frob + ")";
+  }
 }
