+++ extensions/data-transfer/portability-data-transfer-rememberthemilk/src/main/java/org/dataportabilityproject/transfer/rememberthemilk/model/tasks/GetListsResponse.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.transfer.microsoft.model.tasks;
+package org.dataportabilityproject.transfer.rememberthemilk.model.tasks;
 
 import com.google.api.client.util.Key;
 
-public class TaskAddResponse extends RememberTheMilkResponse {
+public class GetListsResponse extends RememberTheMilkResponse {
 
-  @Key("list")
-  public TaskList taskList;
+  @Key("lists")
+  public ListInfoList listInfoList;
 }
