+++ extensions/data-transfer/portability-data-transfer-rememberthemilk/src/main/java/org/dataportabilityproject/transfer/rememberthemilk/model/tasks/Notes.java
@@ -13,16 +13,21 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.transfer.microsoft.model.tasks;
+package org.dataportabilityproject.transfer.rememberthemilk.model.tasks;
 
 import com.google.api.client.util.Key;
+import com.google.common.base.Joiner;
 
-/** A generic response from the remember the milk service. */
-public abstract class RememberTheMilkResponse {
+import java.util.List;
 
-  @Key("@stat")
-  public String stat;
+/** A collection of notes that exist on a task/taskseries */
+public class Notes {
 
-  @Key("err")
-  public Error error;
+  @Key("note")
+  public List<String> notes;
+
+  @Override
+  public String toString() {
+    return (notes == null || notes.isEmpty()) ? "" : Joiner.on("; ").join(notes);
+  }
 }
