+++ extensions/data-transfer/portability-data-transfer-rememberthemilk/src/main/java/org/dataportabilityproject/transfer/rememberthemilk/model/tasks/TaskAddResponse.java
@@ -13,12 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.transfer.microsoft.model.tasks;
+package org.dataportabilityproject.transfer.rememberthemilk.model.tasks;
 
 import com.google.api.client.util.Key;
 
-public class GetListsResponse extends RememberTheMilkResponse {
+public class TaskAddResponse extends RememberTheMilkResponse {
 
-  @Key("lists")
-  public ListInfoList listInfoList;
+  @Key("list")
+  public TaskList taskList;
 }
