+++ extensions/data-transfer/portability-data-transfer-rememberthemilk/src/main/java/org/dataportabilityproject/transfer/rememberthemilk/model/tasks/TimelineCreateResponse.java
@@ -13,26 +13,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.transfer.microsoft.model.tasks;
+package org.dataportabilityproject.transfer.rememberthemilk.model.tasks;
 
 import com.google.api.client.util.Key;
-import com.google.common.base.Joiner;
 
-import java.util.List;
+public class TimelineCreateResponse extends RememberTheMilkResponse {
 
-/** A set of task lists. */
-public class Tasks {
-
-  @Key("@rev")
-  public String rev;
-
-  @Key("list")
-  public List<TaskList> list;
-
-  @Override
-  public String toString() {
-    return String.format(
-        "Tasks(rev=%s List=%s)",
-        rev, (null == list || list.isEmpty()) ? "" : Joiner.on("\n").join(list));
-  }
+  @Key("timeline")
+  public String timeline;
 }
