+++ extensions/data-transfer/portability-data-transfer-rememberthemilk/src/main/java/org/dataportabilityproject/transfer/rememberthemilk/tasks/RememberTheMilkTasksImporter.java
@@ -13,29 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.dataportabilityproject.transfer.microsoft.model.tasks;
 
-import com.google.api.client.util.Key;
-import com.google.common.base.Joiner;
+package org.dataportabilityproject.transfer.rememberthemilk.tasks;
 
-import java.util.List;
+import org.dataportabilityproject.spi.transfer.provider.ImportResult;
+import org.dataportabilityproject.spi.transfer.provider.Importer;
+import org.dataportabilityproject.types.transfer.auth.AuthData;
+import org.dataportabilityproject.types.transfer.models.tasks.TaskContainerResource;
 
-/** A list of one or more {@link Task} contained in a {@link TaskSeries}. */
-public class TaskList {
+import java.util.UUID;
 
-  @Key("@id")
-  public int id;
-
-  @Key("taskseries")
-  public List<TaskSeries> taskSeriesList;
-
+public class RememberTheMilkTasksImporter implements Importer<AuthData, TaskContainerResource> {
   @Override
-  public String toString() {
-    return String.format(
-        "(list id=%d children:[%s])",
-        id,
-        (taskSeriesList == null || taskSeriesList.isEmpty())
-            ? ""
-            : Joiner.on("\n").join(taskSeriesList));
+  public ImportResult importItem(UUID jobId, AuthData authData, TaskContainerResource data) {
+    return null;
   }
 }
